don't start mixnet client for lp reg, with fallback
